renamed and adapted particle container for interface usage
